test(terminal): split session manager coverage
