+++ src/gateway/terminal/session-manager.test.ts
@@ -21,41 +21,6 @@ function deferred<T>() {
 }
 
 describe("TerminalSessionManager", () => {
-  it("seeds only fresh operator sessions and delivers the intro to the opening client", async () => {
-    vi.useFakeTimers();
-    const emit = vi.fn();
-    const manager = new TerminalSessionManager({ emit, spawn: async () => makeFakePty() });
-    try {
-      const operator = await manager.open(baseRequest());
-      if (!operator.ok) {
-        throw new Error("expected operator open");
-      }
-      const intro = composeTerminalIntroBanner(80);
-      expect(manager.snapshot(operator.sessionId)).toBe(intro);
-
-      await vi.advanceTimersByTimeAsync(4);
-      expect(emit).toHaveBeenCalledWith("conn-1", TERMINAL_EVENT_DATA, {
-        sessionId: operator.sessionId,
-        seq: intro.length,
-        data: intro,
-      });
-
-      emit.mockClear();
-      const agent = await manager.open(
-        baseRequest({ owner: { kind: "agent", agentSessionKey: "agent:main:main" } }),
-      );
-      if (!agent.ok) {
-        throw new Error("expected agent open");
-      }
-      expect(manager.snapshotAgent("agent:main:main", agent.sessionId)).toBe("");
-      await vi.advanceTimersByTimeAsync(4);
-      expect(emit).not.toHaveBeenCalled();
-    } finally {
-      manager.disposeAll();
-      vi.useRealTimers();
-    }
-  });
-
   it("kills a backend that finishes after its open request is cancelled", async () => {
     const spawned = deferred<FakeTerminalPty>();
     const controller = new AbortController();
@@ -863,64 +828,6 @@ describe("TerminalSessionManager agent ownership", () => {
     expect(fake.paused).toBe(false);
     expect(fake.resumeCalls).toBeGreaterThanOrEqual(1);
     expect(manager.size).toBe(1);
-  });
-});
-
-describe("TerminalSessionManager output ring", () => {
-  it("bounds buffered output by evicting whole head chunks", async () => {
-    const fake = makeFakePty();
-    const manager = new TerminalSessionManager({
-      emit: vi.fn(),
-      spawn: async () => fake,
-      scrollbackChars: 8,
-    });
-    const outcome = await manager.open(baseRequest());
-    if (!outcome.ok) {
-      throw new Error("expected open");
-    }
-    fake.emitData("abcd");
-    fake.emitData("efgh");
-    expect(manager.snapshot(outcome.sessionId)).toBe("abcdefgh");
-    fake.emitData("ijkl");
-    // Cap exceeded: the oldest whole chunk goes; boundaries stay intact.
-    expect(manager.snapshot(outcome.sessionId)).toBe("efghijkl");
-  });
-
-  it("keeps only the tail of a single oversized chunk", async () => {
-    const fake = makeFakePty();
-    const manager = new TerminalSessionManager({
-      emit: vi.fn(),
-      spawn: async () => fake,
-      scrollbackChars: 8,
-    });
-    const outcome = await manager.open(baseRequest());
-    if (!outcome.ok) {
-      throw new Error("expected open");
-    }
-    fake.emitData("0123456789AB");
-    expect(manager.snapshot(outcome.sessionId)).toBe("456789AB");
-  });
-
-  it("does not retain a leading lone low surrogate from an oversized chunk", async () => {
-    const fake = makeFakePty();
-    const manager = new TerminalSessionManager({
-      emit: vi.fn(),
-      spawn: async () => fake,
-      scrollbackChars: 3,
-    });
-    const outcome = await manager.open(baseRequest());
-    if (!outcome.ok) {
-      throw new Error("expected open");
-    }
-
-    fake.emitData("ab😀cd");
-
-    expect(manager.snapshot(outcome.sessionId)).toBe("cd");
-  });
-
-  it("returns undefined for unknown sessions", () => {
-    const manager = new TerminalSessionManager({ emit: vi.fn() });
-    expect(manager.snapshot("nope")).toBeUndefined();
   });
 });
